Moved color sensor files from launcher folder to intake folder
