+++ src/main/java/frc/robot/subsystems/intake/ColorSensorIO.java
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
 
-package frc.robot.subsystems.launcher;
+package frc.robot.subsystems.intake;
 
 import org.littletonrobotics.junction.AutoLog;
